requirements.txt
dnspython==2.3.0
fastapi==0.98.0
httpx==0.24.1
langchain==0.0.230
load-dotenv==0.1.0
mypy-extensions==1.0.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
psycopg2-binary
pydantic==1.10.9
pymongo==4.4.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==1.4.49
sql-metadata==2.8.0
sqlparse==0.4.4
sshtunnel==0.4.0
uvicorn==0.22.0
pytest== 7.4.0
ipdb==0.13.13
snowflake-connector-python==3.0.4
snowflake-sqlalchemy==1.4.7
databricks-sql-connector==2.7.0
sqlalchemy-databricks==0.2.0
sqlalchemy-bigquery==1.6.1
chromadb==0.3.26
pytest-dotenv==0.5.2
pinecone-client==2.2.2
cryptography==40.0.2
sphinx==6.2.1
sphinx-book-theme==1.0.1
streamlit
